refactor(cli): split mcp execution record tests
